Frontend changes in Landing Page, Venue List, Decor List. Tailwind installed but not yet used
