+++ frontend/src/components/pages/venue/ShimmerUI.js
@@ -1,0 +1,68 @@
+import React from 'react'
+
+const ShimmerUI = () => {
+  return (
+
+    <>
+    <div className="Venuecards-Shimmer styles_container V2__1_U51">
+            <div className='vc-img-Shimmer'>
+            {/* <img src = {image}/> */}
+            </div>
+            
+            <div className='vc-venue-details-Shimmer'>
+                <div className='vc-venue-name-big-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'>
+                   
+                </div>
+                
+            </div>
+            
+            
+            
+            
+        </div>
+        <div className="Venuecards-Shimmer styles_container V2__1_U51">
+            <div className='vc-img-Shimmer'>
+            {/* <img src = {image}/> */}
+            </div>
+            
+            <div className='vc-venue-details '>
+                <div className='vc-venue-name-big-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'>
+                   
+                </div>
+                
+            </div>
+            
+            
+            
+            
+        </div>
+        <div className="Venuecards-Shimmer styles_container V2__1_U51">
+            <div className='vc-img-Shimmer'>
+            
+            </div>
+            
+            <div className='vc-venue-details '>
+                <div className='vc-venue-name-big-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'></div>
+                <div className='vc-venue-name-Shimmer'>
+                   
+                </div>
+                
+            </div>  
+            
+            
+            
+            
+        </div>
+    </>
+  )
+}
+
+export default ShimmerUI
