--- frontend/src/components/venue/venue.css
@@ -1,231 +1,0 @@
-.maindivVenue{
-    font-family: 'SF Pro Display',-apple-system,BlinkMacSystemFont,sans-serif !important;
-}
-
-
-.Listcontainer{
-    display: flex;
-    background-color: #f5f6f8;
-    align-items: center;
-    justify-content: center;
-    margin-top: 40px;
-    width: 100%;
-
-    
-}
-
-.styles_containerV2__1_U51{
-    contain-intrinsic-size: 250px;
-    box-shadow: 0 4px 29px #e8e8e8;
-    background: none;
-}
-
-
-.div-srch{
-    display: flex;
-    background-color: rgb(255, 255, 255);
-    align-items: center;
-    justify-content: center;
-    /* position: relative;
-    padding: 15px 0 0;
-    margin-top: 10px;
-    width: 50%; */
-}
-.div-srch .div-service{
-    display: flex;
-    background-color: rgba(0, 255, 255, 0);
-    margin-right: 30px;
-    justify-content: left;
-}
-
-.div-srch button{
-    background-color: #fde830;
-    min-width: 100px;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    font-weight: 500;
-    font:  1em sans-serif;
-    border: black ;
-    border-radius: 16px;
-    padding: 10px;
-    margin: 5px;
-    margin-bottom: 10px;
-    cursor: pointer;
-
-}
-
-.div-srch button:hover {
-    background-color: #f9ef93;
-    box-shadow: 0px 0px 20px -18px;
-  }
-.div-srch button:active {
-    transform: scale(.97);
-  }
-
-.div-srch input{
-    
-    font-family: inherit;
-    width: 25%;
-    border: 0;
-    border-bottom: 2px solid gray;
-    outline: 0;
-    font-size: 1.3rem;
-    color: rgb(3, 3, 3);
-    padding: 7px 0;
-    background: transparent;
-    transition: border-color 0.2s;
-
-    
-
-    
-}
-
-.Venuecontainer{
-    display: flex;
-    width: 80%;
-    flex-flow: row wrap;
-    justify-content: center;
-    overflow-y: visible;
-    /* flex-direction: column; */
-    background-color: rgba(255, 193, 146, 0);
-}
-
-.Venuecards{
-    display: flex;
-    padding: 15px;
-    justify-content: left;
-    align-items: center;
-    flex-direction: row;
-    background-color: rgb(255, 255, 255);
-    margin: 10px;
-    border-bottom-right-radius: 36px;
-    width: 100%;
-}
-
-.Venuecards:hover{
-    cursor: pointer;
-    
-}
-
-.vc-img img{
-    border-radius: 16px;
-    margin: 10px;
-}
-
-
-.vc-venue-details{
-    background-color: rgb(255, 255, 255);
-    margin: 10px;
-    
-    height: 85%;
-}
-.vc-venue-name{
-    display: flex;
-    font-size: 2em;
-    align-items: start;
-    justify-content: start;
-    /* background-color: rgb(255, 255, 255); */
-}
-
-
-.vc-venue-details button{
-    background-color: #fde830;
-    min-width: 100px;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    font-weight: 500;
-    font:  1em sans-serif;
-    border: black ;
-    border-radius: 16px;
-    padding: 10px;
-    margin: 5px;
-    margin-bottom: 10px;
-    cursor: pointer;
-}
-
-.vc-venue-details button:hover {
-    background-color: #f9ef93;
-    box-shadow: 0px 0px 20px -18px;
-  }
-
-.vc-venue-details button:active {
-    transform: scale(.97);
-  }
-
-
-
-  /* Shimmer UI CSS for Venue List */
-
-  .Venuecontainer-Shimmer{
-    display: flex;
-    width: 80%;
-    flex-flow: row wrap;
-    justify-content: center;
-    overflow-y: visible;
-    /* flex-direction: column; */
-    background-color: rgba(255, 193, 146, 0);
-}
-
-.Venuecards-Shimmer{
-    display: flex;
-    padding: 15px;
-    justify-content: left;
-    align-items: center;
-    flex-direction: row;
-    background-color: rgb(255, 255, 255);
-    margin: 10px;
-    border-bottom-right-radius: 36px;
-    width: 100%;
-    height: 195px;
-}
-
-.vc-img-Shimmer{
-    border-radius: 16px;
-    margin: 10px;
-    width: 150px;
-    height: 150px;
-    background: linear-gradient(to right, #f6f7f8 0%, #edeef1 20%, #f6f7f8 40%, #f6f7f8 100%);
-    background-size: 300%;
-    background-position-x: 100%;
-    animation: shimmer 1s infinite linear;
-}
-@keyframes shimmer {
-    to {
-       background-position-x: 0%
-    }
- }
-
-
-.vc-venue-details-Shimmer{
-    background-color: rgb(255, 255, 255);
-    margin: 10px;
-    
-    height: 85%;
-}
-
-.vc-venue-name-big-Shimmer{
-    background-color: lightgray;
-    width: 240px;
-    display: flex;
-    margin: 15px;
-    height: 30px;
-    align-items: start;
-    justify-content: start;
-    background: linear-gradient(to right, #f6f7f8 0%, #edeef1 20%, #f6f7f8 40%, #f6f7f8 100%);
-    background-size: 300%;
-    background-position-x: 100%;
-    animation: shimmer 1s infinite linear;
-    /* background-color: rgb(255, 255, 255); */
-}
-.vc-venue-name-Shimmer{
-    background-color: lightgray;
-    width: 140px;
-    display: flex;
-    margin: 15px;
-    height: 20px;
-    align-items: start;
-    justify-content: start;
-    background: linear-gradient(to right, #f6f7f8 0%, #edeef1 20%, #f6f7f8 40%, #f6f7f8 100%);
-    background-size: 300%;
-    background-position-x: 100%;
-    animation: shimmer 1s infinite linear;
-    /* background-color: rgb(255, 255, 255); */
-}
