+++ frontend/src/styles/Login.module.css
@@ -1,7 +1,4 @@
-body{
-    margin: 0;
-    background-color: #ffffff;
-}
+
 
 .container {
     display: flex;
